Move HostConcatResultUtil out of unshimmed classes
Don't use ai.rapidsai.cudf package for spark-rapids Scala classes.
Otherwise it is attempted to be loaded by conventional classloader
